+++ gear/form_qc_coordinator/src/python/form_qc_coordinator_app/coordinator.py
@@ -26,7 +26,7 @@ from gear_execution.gear_trigger import CredentialGearConfigs, GearInfo, trigger
 from jobs.job_poll import JobPoll
 from keys.keys import DefaultValues, FieldNames, MetadataKeys, SysErrorCodes
 from outputs.error_logger import update_error_log_and_qc_metadata
-from outputs.error_models import FileError
+from outputs.error_models import FileError, VisitKeys
 from outputs.error_writer import ListErrorWriter
 from outputs.errors import (
     preprocessing_error,
@@ -309,6 +309,8 @@ class QCCoordinator:
         failed_visit: str,
         ptid_key: str,
         date_col_key: str,
+        visitnum_key: str,
+        naccid_key: str,
     ):
         """Update error metadata in the visit files that were not processed due
         to a failure of a previous visit.
@@ -316,8 +318,10 @@ class QCCoordinator:
         Args:
             remaining_visits: visits that were not processed
             failed_visit: name of the failed visit
-            ptid_key: primary key location in file.info
+            ptid_key: ptid field location in file.info
             date_col_key: date field location in file.info
+            visitnum_key: visitnum field location in file.info
+            naccid_key: naccid field location in file.info
         """
         log.info(
             "Visit %s failed, there are %s subsequent visits for this participant.",
@@ -325,11 +329,13 @@ class QCCoordinator:
             len(remaining_visits),
         )
         log.info("Adding error metadata to respective visit files")
+
         while len(remaining_visits) > 0:
             visit = remaining_visits.popleft()
             file_id = visit["file.file_id"]
             visitdate = visit[date_col_key]
             ptid = visit[ptid_key]
+
             try:
                 visit_file = self.__proxy.get_file(file_id)
             except ApiException as error:
@@ -340,7 +346,15 @@ class QCCoordinator:
                     "Error metadata not updated for visit %s", visit["file.name"]
                 )
                 continue
-            error_obj = previous_visit_failed_error(failed_visit)
+            error_obj = previous_visit_failed_error(
+                prev_visit=failed_visit,
+                visit_keys=VisitKeys(
+                    ptid=ptid,
+                    visitnum=visit.get(visitnum_key),
+                    date=visitdate,
+                    naccid=visit.get(naccid_key),
+                ),
+            )
             self.__update_qc_error_metadata(
                 visit_file=visit_file,
                 error_obj=error_obj,
@@ -411,7 +425,8 @@ class QCCoordinator:
         visit_file: FileEntry,
         ptid: str,
         visitdate: str,
-        visitnum: Optional[str],
+        visitnum: Optional[str] = None,
+        naccid: Optional[str] = None,
     ) -> Optional[Dict[str, FileEntry]]:
         """Populate the inputs required for QC gear, report errors if required
         input files cannot be found.
@@ -421,6 +436,7 @@ class QCCoordinator:
             ptid: participant identifier
             visitdate: visit date
             visitnum (optional): visit number
+            naccid (optional): NACCID
 
         Returns:
             Dict[str, FileEntry (optional): gear input dictionary or None
@@ -447,8 +463,9 @@ class QCCoordinator:
                     field=FieldNames.MODULE,
                     value=self.__module,
                     error_code=SysErrorCodes.UDS_NOT_APPROVED,
-                    ptid=ptid,
-                    visitnum=visitnum,
+                    visit_keys=VisitKeys(
+                        ptid=ptid, visitnum=visitnum, date=visitdate, naccid=naccid
+                    ),
                 )
                 self.__update_visit_metadata_on_failure(
                     ptid=ptid,
@@ -480,6 +497,7 @@ class QCCoordinator:
         ptid_key = MetadataKeys.get_column_key(FieldNames.PTID)
         date_col_key = MetadataKeys.get_column_key(self.__module_configs.date_field)
         visitnum_key = MetadataKeys.get_column_key(FieldNames.VISITNUM)
+        naccid_key = MetadataKeys.get_column_key(FieldNames.NACCID)
 
         # sort the visits in the ascending order of visit date
         sorted_visits = sorted(visits, key=lambda d: d[date_col_key])
@@ -492,6 +510,7 @@ class QCCoordinator:
             visitdate = visit[date_col_key]
             ptid = visit[ptid_key]
             visitnum = visit.get(visitnum_key)
+            naccid = visit.get(naccid_key)
 
             # skip if module file was validated after the finalization trigger
             if self.__is_outdated_trigger(visit):
@@ -504,6 +523,7 @@ class QCCoordinator:
                     ptid=ptid,
                     visitdate=visitdate,
                     visitnum=visitnum,
+                    naccid=naccid,
                 )
             except GearExecutionError as error:
                 raise error
@@ -535,7 +555,8 @@ class QCCoordinator:
             # If QC gear did not complete, stop evaluating any subsequent visits
             if not JobPoll.is_job_complete(self.__proxy, job_id):
                 error_obj = system_error(
-                    f"Errors occurred while running gear {gear_name} on this file"
+                    message=f"Errors occurred while running gear {gear_name}",
+                    visit_keys=VisitKeys(ptid=ptid, visitnum=visitnum, date=visitdate),
                 )
                 self.__update_visit_metadata_on_failure(
                     ptid=ptid,
@@ -568,4 +589,6 @@ class QCCoordinator:
                 failed_visit=failed_visit,
                 ptid_key=ptid_key,
                 date_col_key=date_col_key,
+                visitnum_key=visitnum_key,
+                naccid_key=naccid_key,
             )